Add `FileCollectionMixin` and use it
