+++ weblab/datasets/models.py
@@ -2,12 +2,11 @@ from django.core.validators import MinLengthValidator
 from django.db import models
 from django.utils.text import get_valid_filename
 
-from core.combine import ArchiveReader
-from core.models import UserCreatedModelMixin, VisibilityModelMixin
+from core.models import FileCollectionMixin, UserCreatedModelMixin, VisibilityModelMixin
 from entities.models import ProtocolEntity
 
 
-class Dataset(UserCreatedModelMixin, VisibilityModelMixin, models.Model):
+class Dataset(UserCreatedModelMixin, VisibilityModelMixin, FileCollectionMixin, models.Model):
     """Prototyping class for experimental datasets
     """
     name = models.CharField(validators=[MinLengthValidator(2)], max_length=255)
@@ -35,20 +34,6 @@ class Dataset(UserCreatedModelMixin, VisibilityModelMixin, models.Model):
     @property
     def archive_name(self):
         return get_valid_filename(self.name + '.zip')
-
-    @property
-    def archive_path(self):
-        return self.abs_path / self.archive_name
-
-    @property
-    def files(self):
-        if self.archive_path.exists():
-            return ArchiveReader(str(self.archive_path)).files
-        else:
-            return []
-
-    def open_file(self, name):
-        return ArchiveReader(str(self.archive_path)).open_file(name)
 
 
 class DatasetFile(models.Model):
