+++ weblab/experiments/models.py
@@ -4,8 +4,7 @@ from pathlib import Path
 from django.conf import settings
 from django.db import models
 
-from core.combine import ArchiveReader
-from core.models import UserCreatedModelMixin
+from core.models import FileCollectionMixin, UserCreatedModelMixin
 from core.visibility import Visibility, get_joint_visibility, visibility_check
 from entities.models import ModelEntity, ProtocolEntity
 
@@ -119,7 +118,7 @@ class Experiment(UserCreatedModelMixin, models.Model):
             return ''
 
 
-class ExperimentVersion(UserCreatedModelMixin, models.Model):
+class ExperimentVersion(UserCreatedModelMixin, FileCollectionMixin, models.Model):
     STATUS_QUEUED = "QUEUED"
     STATUS_RUNNING = "RUNNING"
     STATUS_SUCCESS = "SUCCESS"
@@ -177,13 +176,9 @@ class ExperimentVersion(UserCreatedModelMixin, models.Model):
     def abs_path(self):
         return Path(settings.EXPERIMENT_BASE, str(self.id))
 
-    def mkdir(self):
-        """Create the folder for this version's results (and parents if needed)."""
-        self.abs_path.mkdir(exist_ok=True, parents=True)
-
     @property
-    def archive_path(self):
-        return self.abs_path / 'results.omex'
+    def archive_name(self):
+        return 'results.omex'
 
     @property
     def signature(self):
@@ -208,16 +203,6 @@ class ExperimentVersion(UserCreatedModelMixin, models.Model):
         self.status = status
         self.return_text = txt
         self.save()
-
-    @property
-    def files(self):
-        if self.archive_path.exists():
-            return ArchiveReader(str(self.archive_path)).files
-        else:
-            return []
-
-    def open_file(self, name):
-        return ArchiveReader(str(self.archive_path)).open_file(name)
 
 
 class RunningExperiment(models.Model):
